+++ src/realm/util/file_mapper.hpp
@@ -72,6 +72,14 @@ inline void set_page_reclaim_governor_to_default()
     set_page_reclaim_governor(nullptr);
 }
 
+// There are several globals that rely on being process specific.
+// The unit tests which use fork() should start with empty mappings.
+// This also resets the page reclaimer thread because if fork happens while
+// running with the global mutex `mapping_mutex` locked, the child process
+// will hang due to that mutex never being unlocked.
+// We do not support fork() in production.
+void clear_mappings_before_test_forks();
+
 // Retrieves the number of in memory decrypted pages, across all open files.
 size_t get_num_decrypted_pages();