+++ test/object-store/sync/flx_sync.cpp
@@ -256,6 +256,32 @@ TEST_CASE("flx: connect to FLX-enabled app", "[sync][flx][app]") {
     });
 }
 
+TEST_CASE("flx: test commands work") {
+    FLXSyncTestHarness harness("test_commands");
+    harness.do_with_new_realm([&](const SharedRealm& realm) {
+        wait_for_upload(*realm);
+        nlohmann::json command_request = {
+            {"command", "PAUSE_ROUTER_SESSION"},
+        };
+        auto resp_body =
+            SyncSession::OnlyForTesting::send_test_command(*realm->sync_session(), command_request.dump()).get();
+        REQUIRE(resp_body == "{}");
+
+        auto bad_status =
+            SyncSession::OnlyForTesting::send_test_command(*realm->sync_session(), "foobar: }").get_no_throw();
+        REQUIRE(bad_status.get_status() == ErrorCodes::LogicError);
+        REQUIRE_THAT(bad_status.get_status().reason(),
+                     Catch::Matchers::ContainsSubstring("Invalid json input to send_test_command"));
+
+        bad_status =
+            SyncSession::OnlyForTesting::send_test_command(*realm->sync_session(), "{\"cmd\": \"\"}").get_no_throw();
+        REQUIRE_FALSE(bad_status.is_ok());
+        REQUIRE(bad_status.get_status() == ErrorCodes::LogicError);
+        REQUIRE(bad_status.get_status().reason() ==
+                "Must supply command name in \"command\" field of test command json object");
+    });
+}
+
 static auto make_error_handler()
 {
     auto [error_promise, error_future] = util::make_promise_future<SyncError>();